Fixes DCV - datatype bug

+++ src/app/model/rdf/sparql/datacube/package.scala
@@ -1,6 +1,6 @@
 package model.rdf.sparql
 
 package object datacube {
-  type MeasuresByKey = Map[String, Option[Long]]
+  type MeasuresByKey = Map[String, Option[BigDecimal]]
   type SlicesByKey = Map[String, MeasuresByKey]
 }
